+++ src/client/java/io/github/tr100000/trutils/ClientDelegateClientImpl.java
@@ -3,7 +3,9 @@ package io.github.tr100000.trutils;
 import io.github.tr100000.trutils.api.gui.component.AbstractComponentScreenHandler;
 import io.github.tr100000.trutils.api.gui.component.ClientComponentScreenHandler;
 import net.minecraft.world.entity.player.Inventory;
+import org.jspecify.annotations.NullMarked;
 
+@NullMarked
 public class ClientDelegateClientImpl extends ClientDelegate {
     @Override
     public boolean isClient() {