+++ src/main/java/io/github/tr100000/trutils/api/datagen/GeneratedPackResources.java
@@ -6,6 +6,7 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.server.packs.PackType;
 import net.minecraft.server.packs.PathPackResources;
 import net.minecraft.server.packs.resources.IoSupplier;
+import org.jetbrains.annotations.ApiStatus;
 import org.jspecify.annotations.Nullable;
 
 import java.io.ByteArrayInputStream;
@@ -14,6 +15,7 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.Optional;
 
+@ApiStatus.Internal
 public class GeneratedPackResources extends PathPackResources {
     private final ModContainer mod;
     private final PackType type;