+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 **Updated to 26.2**
 
 ### Changed
-- Changed many things in `RegistryHelper`, `ItemRegistryHelper`, and `BlockRegistryHelper` because the game needs `ResourceKey` everywhere
+- Changed many things in `RegistryHelper`, `ItemRegistryHelper`, and `BlockRegistryHelper` because the game needs `ResourceKey` everywhere now