+++ CHANGELOG.md
@@ -1,3 +1,2 @@
-### Added
-- Added the `trutils:datagen/client` entrypoint for client-only runtime datagen
-- Mods can now have multiple runtime datagen entrypoints
+### Changed
+- Messed with runtime datagen internals to closer match normal fabric data generation behavior